Disable `yas-field-highlight-face`
It tends to show up in expansions having a single argument ($1, and no $2), which is redundant and a bit annoying, since the highlight will not go away by pressing SPC.

As a side-effect, expansions having $2 or more will have no highlighting, which will be bit of a more opaque UX. But it still works, e.g. pressing TAB will keep moving to the next tabstop.

One could also request upstream that no highlights are placed if there's just $1.